Return stderr and return codes to libvirtd
Hooks are now processed in a loop that watches return codes from each; the first (if any) to return a non-zero return code inspires this script to re-emit any stderr output captured, and then exit with a matching return code. The simplest case in which this is valuable, perhaps, is a prepare/begin script that wants to tell libvirtd not to proceed with the starting of a domain.